+++ sample/src/commonMain/kotlin/com/github/panpf/sketch/sample/ui/gallery/PhotoPagerScreen.common.kt
@@ -112,7 +112,6 @@ class PhotoPagerScreen(private val params: PhotoPagerParams) : BaseScreen() {
 
             HorizontalPager(
                 state = pagerState,
-                beyondBoundsPageCount = 0,
                 modifier = Modifier.fillMaxSize()
             ) { index ->
                 val pageSelected by remember {