+++ sample/src/desktopMain/kotlin/com/github/panpf/sketch/sample/ui/test/FetcherTestScreen.desktop.kt
@@ -1,7 +1,5 @@
 package com.github.panpf.sketch.sample.ui.test
 
-import androidx.compose.ui.ExperimentalComposeUiApi
-import androidx.compose.ui.res.ResourceLoader
 import com.github.panpf.sketch.PlatformContext
 import com.github.panpf.sketch.fetch.newComposeResourceUri
 import com.github.panpf.sketch.fetch.newFileUri
@@ -9,8 +7,8 @@ import com.github.panpf.sketch.fetch.newKotlinResourceUri
 import com.github.panpf.sketch.images.Base64Images
 import com.github.panpf.sketch.images.HttpImages
 import com.github.panpf.sketch.images.ImageFile
-import com.github.panpf.sketch.images.ResourceImages
 import com.github.panpf.sketch.images.ResourceImageFile
+import com.github.panpf.sketch.images.ResourceImages
 import com.github.panpf.sketch.sample.appId
 import com.github.panpf.sketch.sample.resources.Res
 import kotlinx.coroutines.Dispatchers
@@ -18,6 +16,7 @@ import kotlinx.coroutines.withContext
 import net.harawata.appdirs.AppDirsFactory
 import org.jetbrains.compose.resources.ExperimentalResourceApi
 import java.io.File
+import java.io.InputStream
 
 @OptIn(ExperimentalResourceApi::class)
 actual suspend fun buildFetcherTestItems(
@@ -42,7 +41,6 @@ actual suspend fun buildFetcherTestItems(
     }
 }
 
-@OptIn(ExperimentalComposeUiApi::class)
 private suspend fun getFileUriTestFile(image: ImageFile): File =
     withContext(Dispatchers.IO) {
         val appDataDir = AppDirsFactory.getInstance().getUserDataDir(
@@ -54,7 +52,7 @@ private suspend fun getFileUriTestFile(image: ImageFile): File =
         val imageFile = File(appDataDir, resourceImage.resourceName)
         if (!imageFile.exists()) {
             imageFile.parentFile.mkdirs()
-            ResourceLoader.Default.load(resourceImage.resourceName).use { input ->
+            ClassLoaderResourceLoader.Default.load(resourceImage.resourceName).use { input ->
                 imageFile.outputStream().use { output ->
                     input.copyTo(output)
                 }
@@ -62,3 +60,23 @@ private suspend fun getFileUriTestFile(image: ImageFile): File =
         }
         imageFile
     }
+
+
+/**
+ * Resource loader based on JVM current context class loader.
+ *
+ * Copy from compose Resources.kt
+ */
+private class ClassLoaderResourceLoader {
+
+    companion object {
+        val Default = ClassLoaderResourceLoader()
+    }
+
+    fun load(resourcePath: String): InputStream {
+        val contextClassLoader = Thread.currentThread().contextClassLoader!!
+        val resource = contextClassLoader.getResourceAsStream(resourcePath)
+            ?: (::ClassLoaderResourceLoader.javaClass).getResourceAsStream(resourcePath)
+        return requireNotNull(resource) { "Resource $resourcePath not found" }
+    }
+}